requirements.txt
Django~=2.1.5
Pillow~=5.4.1
numpy~=1.16.0
chainer~=5.1.0